re #4, factor getInputType helper from getframe_*. use VideoWriter in R2014b and later (avifile deprecated). fix typos.

+++ @myavifile/addframe.m
@@ -335,7 +335,7 @@ return;
 
 % ------------------------------------------------------------------------
 function inputType = getInputType(frame)
-  if isscalar(frame) && ishandle(frame) && (frame > 0)
+  if isscalar(frame) && ishandle(frame) && (double(frame) > 0)
     inputType = get(frame,'type');
   elseif isstruct(frame) && isfield(frame,'cdata')
     inputType = 'movie';
